refactor: 출석 상태 테스트를 AttendanceStatusTest로 이동
