+++ src/test/java/domain/AttendanceTest.java
@@ -1,6 +1,5 @@
 package domain;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
@@ -52,116 +51,6 @@ class AttendanceTest {
 
         assertThat(localDate).isEqualTo(LocalDate.of(2025, 2, 27));
         assertThat(localTime).isEqualTo(LocalTime.of(10, 6));
-    }
-
-    @Test
-    void 월요일_출석_테스트() {
-        Attendance attendance = new Attendance(
-                LocalDateTime.of(2024, 12, 16, 13, 0)
-        );
-
-        AttendanceStatus attendanceStatus = attendance.judge();
-
-        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("출석");
-    }
-
-    @Test
-    void 월요일_출석_경계값_테스트() {
-        Attendance attendance = new Attendance(
-                LocalDateTime.of(2024, 12, 16, 13, 5)
-        );
-
-        AttendanceStatus attendanceStatus = attendance.judge();
-
-        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("출석");
-    }
-
-    @Test
-    void 월요일_지각_테스트() {
-        Attendance attendance = new Attendance(
-                LocalDateTime.of(2024, 12, 16, 13, 6)
-        );
-
-        AttendanceStatus attendanceStatus = attendance.judge();
-
-        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("지각");
-    }
-
-    @Test
-    void 월요일_지각_경계값_테스트() {
-        Attendance attendance = new Attendance(
-                LocalDateTime.of(2024, 12, 16, 13, 30)
-        );
-
-        AttendanceStatus attendanceStatus = attendance.judge();
-
-        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("지각");
-    }
-
-    @Test
-    void 월요일_결석_테스트() {
-        Attendance attendance = new Attendance(
-                LocalDateTime.of(2024, 12, 16, 13, 31)
-        );
-
-        AttendanceStatus attendanceStatus = attendance.judge();
-
-        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("결석");
-    }
-
-    @Test
-    void 화요일_출석_테스트() {
-        Attendance attendance = new Attendance(
-                LocalDateTime.of(2024, 12, 17, 10, 0)
-        );
-
-        AttendanceStatus attendanceStatus = attendance.judge();
-
-        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("출석");
-    }
-
-    @Test
-    void 화요일_출석_경계값_테스트() {
-        Attendance attendance = new Attendance(
-                LocalDateTime.of(2024, 12, 17, 10, 5)
-        );
-
-        AttendanceStatus attendanceStatus = attendance.judge();
-
-        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("출석");
-    }
-
-    @Test
-    void 화요일_지각_테스트() {
-        Attendance attendance = new Attendance(
-                LocalDateTime.of(2024, 12, 17, 10, 6)
-        );
-
-        AttendanceStatus attendanceStatus = attendance.judge();
-
-        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("지각");
-    }
-
-    @Test
-    void 화요일_지각_경계값_테스트() {
-        Attendance attendance = new Attendance(
-                LocalDateTime.of(2024, 12, 17, 10, 30)
-        );
-
-        AttendanceStatus attendanceStatus = attendance.judge();
-
-        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("지각");
-    }
-
-    @Test
-    void 화요일_결석_테스트() {
-        Attendance attendance = new Attendance(
-                LocalDateTime.of(2024, 12, 17, 10, 31)
-        );
-
-        AttendanceStatus attendanceStatus = attendance.judge();
-
-        Assertions.assertThat(attendanceStatus.getName()).isEqualTo("결석");
     }
 
     @Test
